backend/bank_django_nuxt3_api/requirements.txt
Django==4.0.4
django-environ==0.8.1
psycopg2-binary==2.9.3
djangorestframework==3.13.1
django-filter==21.1
markdown==3.3.6
django-cors-headers==3.11.0
gunicorn==20.1.0
factory-boy==3.2.1